Move SDL/OpenGL intialization and window creation into game_main.cpp
