+++ src/game_main.cpp
@@ -22,7 +22,9 @@
 #include "data/game_traits.hpp"
 #include "engine/timing.hpp"
 #include "loader/duke_script_loader.hpp"
+#include "renderer/opengl.hpp"
 #include "sdl_utils/error.hpp"
+#include "sdl_utils/ptr.hpp"
 #include "ui/imgui_integration.hpp"
 
 #include "game_session_mode.hpp"
@@ -31,6 +33,7 @@
 
 RIGEL_DISABLE_WARNINGS
 #include <imgui.h>
+#include <SDL.h>
 RIGEL_RESTORE_WARNINGS
 
 #include <cassert>
@@ -44,6 +47,74 @@ using namespace sdl_utils;
 using RenderTargetBinder = renderer::RenderTargetTexture::Binder;
 
 namespace {
+
+#if defined( __APPLE__) || defined(RIGEL_USE_GL_ES)
+  const auto WINDOW_FLAGS = SDL_WINDOW_FULLSCREEN;
+#else
+  const auto WINDOW_FLAGS = SDL_WINDOW_BORDERLESS;
+#endif
+
+
+// Default values for screen resolution in case we can't figure out the
+// current Desktop size
+const auto DEFAULT_RESOLUTION_X = 1920;
+const auto DEFAULT_RESOLUTION_Y = 1080;
+
+
+template <typename Callback>
+class CallOnDestruction {
+public:
+  explicit CallOnDestruction(Callback&& callback)
+    : mCallback(std::forward<Callback>(callback))
+  {
+  }
+
+  ~CallOnDestruction() {
+    mCallback();
+  }
+
+  CallOnDestruction(const CallOnDestruction&) = delete;
+  CallOnDestruction& operator=(const CallOnDestruction&) = delete;
+
+private:
+  Callback mCallback;
+};
+
+
+template <typename Callback>
+[[nodiscard]] auto defer(Callback&& callback) {
+  return CallOnDestruction{std::forward<Callback>(callback)};
+}
+
+
+SDL_Window* createWindow() {
+  SDL_DisplayMode displayMode;
+  if (SDL_GetDesktopDisplayMode(0, &displayMode) != 0) {
+    displayMode.w = DEFAULT_RESOLUTION_X;
+    displayMode.h = DEFAULT_RESOLUTION_Y;
+  }
+
+#ifdef RIGEL_USE_GL_ES
+  SDL_GL_SetAttribute(SDL_GL_CONTEXT_PROFILE_MASK, SDL_GL_CONTEXT_PROFILE_ES);
+  SDL_GL_SetAttribute(SDL_GL_CONTEXT_MAJOR_VERSION, 2);
+  SDL_GL_SetAttribute(SDL_GL_CONTEXT_MINOR_VERSION, 0);
+#else
+  SDL_GL_SetAttribute(SDL_GL_CONTEXT_PROFILE_MASK, SDL_GL_CONTEXT_PROFILE_CORE);
+  SDL_GL_SetAttribute(SDL_GL_CONTEXT_MAJOR_VERSION, 3);
+  SDL_GL_SetAttribute(SDL_GL_CONTEXT_MINOR_VERSION, 0);
+#endif
+  SDL_GL_SetAttribute(SDL_GL_DOUBLEBUFFER, 1);
+
+  return sdl_utils::check(SDL_CreateWindow(
+    "Rigel Engine",
+    SDL_WINDOWPOS_UNDEFINED,
+    SDL_WINDOWPOS_UNDEFINED,
+    displayMode.w,
+    displayMode.h,
+    WINDOW_FLAGS | SDL_WINDOW_OPENGL));
+}
+
+
 
 struct NullGameMode : public GameMode {
   void handleEvent(const SDL_Event&) override {}
@@ -96,10 +167,33 @@ auto loadScripts(const loader::ResourceLoader& resources) {
 }
 
 
-void gameMain(const StartupOptions& options, SDL_Window* pWindow) {
+void gameMain(const StartupOptions& options) {
+#ifdef _WIN32
+  SDL_setenv("SDL_AUDIODRIVER", "directsound", true);
+#endif
+
+  sdl_utils::check(SDL_Init(SDL_INIT_VIDEO | SDL_INIT_AUDIO));
+  auto sdlGuard = defer([]() { SDL_Quit(); });
+
+  sdl_utils::check(SDL_GL_LoadLibrary(nullptr));
+
+  sdl_utils::Ptr<SDL_Window> pWindow(createWindow());
+  SDL_GLContext pGlContext =
+    sdl_utils::check(SDL_GL_CreateContext(pWindow.get()));
+  auto glGuard = defer([pGlContext]() { SDL_GL_DeleteContext(pGlContext); });
+
+  renderer::loadGlFunctions();
+
+  SDL_DisableScreenSaver();
+  SDL_ShowCursor(SDL_DISABLE);
+
+  ui::imgui_integration::init(
+    pWindow.get(), pGlContext, createOrGetPreferencesPath());
+  auto imGuiGuard = defer([]() { ui::imgui_integration::shutdown(); });
+
   auto userProfile = loadOrCreateUserProfile(options.mGamePath);
 
-  Game game(options.mGamePath, &userProfile, pWindow);
+  Game game(options.mGamePath, &userProfile, pWindow.get());
   game.run(options);
 }
